Fix a type bug in FedAtt when training on CIFAR10

+++ examples/fedatt/fedatt_server.py
@@ -45,7 +45,8 @@ class Server(fedavg.Server):
         for name in baseline_weights.keys():
             atts[name] = self.trainer.zeros(len(weights_received))
             for i, update in enumerate(weights_received):
-                delta = update[name]
+                # convert potential LongTensor to FloatTensor for linalg.norm
+                delta = update[name].type(torch.FloatTensor)
                 atts[name][i] = torch.linalg.norm(-delta)
 
         for name in baseline_weights.keys():
